用户ID使用全局唯一ID

+++ src/main/java/ewing/demo/ThymeleafUser.java
@@ -12,6 +12,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 
+import java.math.BigInteger;
+
 /**
  * Thymeleaf测试控制器。
  *
@@ -42,7 +44,7 @@ public class ThymeleafUser {
 
     @GetMapping("deleteUser")
     @PreAuthorize("hasRole('ROLE_USER')")
-    public ModelAndView deleteUser(Long userId) {
+    public ModelAndView deleteUser(BigInteger userId) {
         userService.deleteUser(userId);
         return index();
     }
