+++ src/main/java/ewing/user/UserService.java
@@ -7,6 +7,7 @@ import ewing.entity.User;
 import ewing.security.RoleAsAuthority;
 import ewing.security.SecurityUser;
 
+import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -16,17 +17,17 @@ public interface UserService {
 
     User addUser(User user);
 
-    User getUser(Long userId);
+    User getUser(BigInteger userId);
 
     User updateUser(User user);
 
     Pages<User> findUsers(Pager pager, String name, String roleName);
 
-    void deleteUser(Long userId);
+    void deleteUser(BigInteger userId);
 
     SecurityUser getByName(String name);
 
-    List<RoleAsAuthority> getUserRoles(Long userId);
+    List<RoleAsAuthority> getUserRoles(BigInteger userId);
 
-    List<Permission> getUserPermissions(Long userId);
+    List<Permission> getUserPermissions(BigInteger userId);
 }
